Updated buttons resolution
