--- Company.java
@@ -1,7 +1,0 @@
-import java.util.ArrayList;
-
-public class Company {
-    private Client cliente;
-
-
-}
